Consolidate TestUtils
